requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
redis==5.2.1
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.20
pydantic-settings==2.8.0
imapclient==3.0.1
openai==1.63.0
email-validator==2.2.0
python-dateutil==2.9.0.post0
jsonschema==4.23.0
cryptography==44.0.1
